refactor: move pre-generation helpers to prompts dir
