+++ api/app/clients/OpenAIClient.js
@@ -2,8 +2,7 @@ const BaseClient = require('./BaseClient');
 const ChatGPTClient = require('./ChatGPTClient');
 const { encoding_for_model: encodingForModel, get_encoding: getEncoding } = require('tiktoken');
 const { maxTokensMap, genAzureChatCompletion } = require('../../utils');
-const { formatMessage } = require('./output_parsers');
-const { truncateText } = require('./prompts');
+const { truncateText, formatMessage } = require('./prompts');
 const { runTitleChain } = require('./chains');
 const { createLLM } = require('./llm');
