+++ api/app/clients/output_parsers/index.js
@@ -1,9 +1,7 @@
 const addImages = require('./addImages');
-const handleInputs = require('./handleInputs');
 const handleOutputs = require('./handleOutputs');
 
 module.exports = {
   addImages,
-  ...handleInputs,
   ...handleOutputs,
 };
