+++ api/app/clients/prompts/index.js
@@ -1,10 +1,14 @@
+const formatMessage = require('./formatMessage');
+const refinePrompts = require('./refinePrompts');
+const handleInputs = require('./handleInputs');
 const instructions = require('./instructions');
 const titlePrompts = require('./titlePrompts');
-const refinePrompts = require('./refinePrompts');
 const truncateText = require('./truncateText');
 
 module.exports = {
+  formatMessage,
   ...refinePrompts,
+  ...handleInputs,
   ...instructions,
   ...titlePrompts,
   truncateText,
